+++ packages/blob-storage-models/src/models/api/IBlobStorageGetResponse.ts
@@ -1,6 +1,6 @@
 // Copyright 2024 IOTA Stiftung.
 // SPDX-License-Identifier: Apache-2.0.
-import type { IJsonLdNodeObject } from "@gtsc/data-json-ld";
+import type { IJsonLdNodeObject } from "@twin.org/data-json-ld";
 
 /**
  * Response to get an entry from blob storage.