+++ packages/blob-storage-service/docs/changelog.md
@@ -1,5 +1,5 @@
-# @gtsc/blob-storage-service - Changelog
+# @twin.org/blob-storage-service - Changelog
 
-## v0.0.3
+## v0.0.1-next.3
 
 - Initial Release